refactor(cache): cache repo root path
