+++ src/lib/git-refs/index.ts
@@ -1,10 +1,8 @@
 import { getRef, otherBranchesWithSameCommit } from "./branch_ref";
 import { getBranchChildrenOrParentsFromGit } from "./branch_relations";
-import cache from "./cache";
 
 export {
   getBranchChildrenOrParentsFromGit,
   getRef,
   otherBranchesWithSameCommit,
-  cache,
 };
