+++ src/lib/utils/repo_root_path.ts
@@ -1,0 +1,24 @@
+import { gpExecSync } from "../../lib/utils/exec_sync";
+import { cache } from "../config";
+import { PreconditionsFailedError } from "../errors";
+export function getRepoRootPath(): string {
+  const cachedRepoRootPath = cache.getRepoRootPath();
+  if (cachedRepoRootPath) {
+    return cachedRepoRootPath;
+  }
+  const repoRootPath = gpExecSync(
+    {
+      command: `git rev-parse --git-dir`,
+    },
+    () => {
+      return Buffer.alloc(0);
+    }
+  )
+    .toString()
+    .trim();
+  if (!repoRootPath || repoRootPath.length === 0) {
+    throw new PreconditionsFailedError("No .git repository found.");
+  }
+  cache.setRepoRootPath(repoRootPath);
+  return repoRootPath;
+}
